refactor: Upgrade to parcel v2 and node 19
