--- .postcssrc
@@ -1,3 +1,0 @@
-{
-  "modules": true
-}
