--- postcss.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  modules: true,
-  sourceMap: true,
-};
