+++ src/content/components/BugList/BugList.module.scss
