+++ src/content/components/BugListView/BugListView.module.scss
