+++ src/content/components/CompletionBar/CompletionBar.module.scss
