+++ src/content/components/Loader/Loader.module.scss
@@ -5,17 +5,13 @@
   margin: 88px auto;
   position: relative;
   font-size: 11px;
-  -webkit-transform: translateZ(0);
-  -ms-transform: translateZ(0);
   transform: translateZ(0);
-  -webkit-animation-delay: -0.16s;
   animation-delay: -0.16s;
   &,
   &::before,
   &::after {
     background: var(--loader-color);
-    -webkit-animation: load1 1s infinite ease-in-out;
-    animation: load1 1s infinite ease-in-out;
+    animation: 1s ease-in-out infinite load1;
     width: 1em;
     height: 4em;
   }
@@ -27,24 +23,14 @@
   }
   &::before {
     left: -1.5em;
-    -webkit-animation-delay: -0.32s;
     animation-delay: -0.32s;
   }
   &::after {
     left: 1.5em;
   }
 }
-@-webkit-keyframes load1 {
-  0%,
-  80%,
-  100% {
-    box-shadow: 0 0;
-    height: 4em;
-  }
-  40% {
-    box-shadow: 0 -2em;
-    height: 5em;
-  }
+.loaderContainer {
+  animation: 1s fadein1;
 }
 @keyframes load1 {
   0%,
@@ -57,9 +43,6 @@
     box-shadow: 0 -2em;
     height: 5em;
   }
-}
-.loaderContainer {
-  animation: fadein1 1s;
 }
 @keyframes fadein1 {
   from {
