+++ src/content/components/Preferences/Preferences.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors";
+@import "../../styles/colors.module";
 
 .container {
   max-width: 900px;
