+++ src/content/components/ReleaseReport/ReleaseReport.module.scss
@@ -41,6 +41,10 @@
   border-top: 1px solid rgba(0,0,0,0.2);
 }
 
+.prevUrlButton {
+  float: left;
+}
+
 .nextUrlButton {
   float: right;
 }
