+++ src/content/components/Triage/Triage.module.scss
