+++ src/content/components/ui/Tabs/Tabs.module.scss
