+++ src/content/index.html
@@ -5,9 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="shortcut icon" href="./img/favicon.ico" />
-  <link rel="apple-touch-icon-precomposed" href="./img/favicon.png">
-  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-  <link rel="stylesheet" href="./index.scss">
+  <link rel="apple-touch-icon-precomposed" href="./img/favicon.png" />
+  <link rel="stylesheet" href="index.scss" />
   <title>Bugzy</title>
   <style>
     body,
@@ -23,6 +22,6 @@
 <body>
   <header id="draggy"></header>
   <main id="root"></main>
-  <script src="./index.tsx"></script>
+  <script src="index.tsx" type="module"></script>
 </body>
 </html>
