+++ src/content/index.scss
@@ -1,11 +1,14 @@
-*, *:before, *:after {
+*,
+*:before,
+*:after {
   box-sizing: border-box;
 }
 
 body {
   margin: 0;
   padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Ubuntu', 'Helvetica Neue', sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Ubuntu",
+    "Helvetica Neue", sans-serif;
   font-size: 13px;
   color: #333;
   -webkit-font-smoothing: antialiased;
@@ -49,18 +52,19 @@ h3 {
 
 code {
   color: #e83e8c;
-  font-family: SFMono-Regular,Menlo,Monaco,Consolas,"Liberation Mono","Courier New",monospace;
+  font-family: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono",
+    "Courier New", monospace;
 }
 
 ::-webkit-scrollbar {
   background-color: transparent;
-  width: .8em;
+  width: 0.8em;
   opacity: 1;
 }
 
 ::-webkit-scrollbar-thumb:window-inactive,
 ::-webkit-scrollbar-thumb {
-  border-radius: .8em;
+  border-radius: 0.8em;
   background: #e5e5e5;
 }
