+++ src/content/styles/gStyles.module.scss
