--- tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "jsx": "react",
-    "module": "commonjs",
-    "allowSyntheticDefaultImports": true,
-    "target": "esnext",
-    "outDir": "./dist",
-    "allowJs": true,
-    "resolveJsonModule": true
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
