--- tsconfig.server.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": [
-    "src/content/**/*",
-    "node_modules/"
-  ]
-}
